requirements.txt
celery==5.1.2
certifi==2018.10.15
chardet==3.0.4
channels==2.4.0
channels-redis==2.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dataclasses==0.6
Django==2.2
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
django-redis==4.11.0
djangorestframework==3.9.4
drf-yasg==1.16.0
idna==2.7
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mypy==0.670
mypy-extensions==0.4.1
openapi-codec==1.3.2
Pillow==5.2.0
psycopg2-binary==2.7.6.1
pysolr==3.8.1
pytz==2018.7
redis==3.3.8
requests==2.20.0
ruamel.yaml==0.15.98
sentry-sdk==0.11.2
simplejson==3.16.0
six==1.12.0
tblib==1.3.2
typed-ast==1.3.1
typing-extensions==3.7.2
uritemplate==3.0.0
urllib3==1.24.1
xlutils==2.0.0
pyyaml==5.1.2
pika==1.1.0
pandas==0.25.1
numpy==1.17.2
django_elasticsearch_dsl==7.1.4
vine==5.0.0
pypinyin==0.41.0
DingtalkChatbot==1.5.2
openpyxl==3.0.9
xlrd==1.2.0